# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: NeNe Invoice API
  version: 0.1.0
  description: >
    NeNe Invoice API contract — self-hosted estimate / invoice / payment
    management for Japanese SMBs (qualified-invoice / 適格請求書 compliant).
    Multi-tenant (organization-scoped); money is always integer cents and tax
    rates are basis points (1000 = 10%). All mutating operations are audited.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /admin/templates/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      tags: [Templates]
      summary: Get template by id
      operationId: getTemplateById
      responses:
        '200':
          description: Template (with line presets)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientCapability'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Templates]
      summary: Update template
      operationId: updateTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientCapability'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    delete:
      tags: [Templates]
      summary: Delete template
      operationId: deleteTemplate
      responses:
        '204':
          description: Template deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientCapability'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdPathParam:
      name: id
      in: path
      required: true
      description: Resource identifier.
      schema:
        type: integer
        minimum: 1
  responses:
    Unauthorized:
      description: Missing or invalid bearer token, or bad credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    InsufficientCapability:
      description: Authenticated but lacking the required capability.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found (or not in the caller's organization).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    ValidationFailed:
      description: Request body or field validation error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      description: RFC 9457 problem details.
      properties:
        type:
          type: string
          description: Problem type URI (base https://nene-invoice.dev/problems/).
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
      required: [type, title, status]
...
    Template:
      type: object
      description: A named quote/invoice template (雛形) with its line presets.
      properties:
        id:
          type: integer
        organization_id:
          type: integer
        name:
          type: string
        notes:
          type: [string, 'null']
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        created_at:
          type: [string, 'null']
        updated_at:
          type: [string, 'null']
      required: [id, organization_id, name, line_items]
...
    UpdateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        notes:
          type: [string, 'null']
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItemInput'
      required: [name]
    LineItem:
      type: object
      description: A line on a quote or invoice. Money is integer cents.
      properties:
        id:
          type: integer
        description:
          type: string
        quantity:
          type: integer
        unit_price_cents:
          type: integer
          description: Unit price in integer cents (smallest currency unit).
        tax_rate_bps:
          type: integer
          description: Tax rate in basis points (1000 = 10%, 800 = 8%).
        sort_order:
          type: integer
        line_subtotal_cents:
          type: integer
          description: quantity × unit_price_cents (pre-tax, not rounded).
      required: [description, quantity, unit_price_cents, tax_rate_bps]
    LineItemInput:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price_cents:
          type: integer
          minimum: 0
        tax_rate_bps:
          type: integer
          enum: [800, 1000]
      required: [description, quantity, unit_price_cents, tax_rate_bps]
...
security:
  - bearerAuth: []